# WIP: add a faster filter to render smart quotes