`Proof-of-concept`

Introduces a filter that uses [RubyPants](https://github.com/jmcnevin/rubypants) to render smart-quotes.
The included benchmark script showed this version to be `~5x` faster on my system

```
Warming up --------------------------------------
            smartify     7.000  i/100ms
         smartypants    38.000  i/100ms
Calculating -------------------------------------
            smartify     75.026  (A± 1.3%) i/s -    378.000  in   5.039730s
         smartypants    378.489  (A± 6.3%) i/s -      1.900k in   5.042268s

Comparison:
         smartypants:      378.5 i/s
            smartify:       75.0 i/s - 5.04x  slower
```

/cc @agriffis @pathawks