@agriffis 2018-04-04
This looks great @ashmaroli 

I'd be interested in two things:

1. Ability to change the options in `_config.yml`

2. Consider a different set of defaults. (This could get into bike shedding, but hopefully useful to consider before merging.) My preferred defaults are: `[1, :named_entities, :prevent_breaks]` (the defaults in rubypants, along with the integer options, are historical at this point)
    * Typographically en dashes aren't as useful as em dashes, they're barely different from hyphens. So it makes sense to render both `--` and `---` as em dashes by default.
    * I don't care much about `:named_entities` versus `:character_entities`, so either way.
    * It's a good thing typographically to `:prevent_breaks`.