@ashmaroli 2018-04-05
@agriffis Thank you very much for your suggestions.
I intentionally kept *configurability* out of the loop so that the filter remains being a *slim & faster alternative* for those users who're using the `smartify` filter to convert just "dumb quotes" and "dashes" to typographic alternatives.

Once the Jekyll Core team expresses interest in merging this (esp. since RubyPants does not convert `>`, `<` , `>>`, `<<`), I'will consider fleshing this out to a utilize a *new* `Converter` class under the hood.

The current set of `*options` were chosen since they were *more in sync* with results obtained from the `smartify` filter.